feat: re-use filter functions
