+++ src/helpers/index.ts
@@ -1,4 +1,5 @@
 // !! helpers are not exported outside of this lib, as they are not used in the app, only in the containers
 
 export * from './components'
+export * from './functions'
 export * from './hooks'
